+++ helpo-web/src/_navVoluntario.js
@@ -1,0 +1,37 @@
+  export default {
+  items: [
+    {
+      name: 'Dashboard',
+      url: '/dashboard',
+      icon: 'icon-speedometer'
+    },
+    {
+      title: true,
+      name: 'Actividades sociales',
+      wrapper: {            // optional wrapper object
+        element: '',        // required valid HTML5 element tag
+        attributes: {}        // optional valid JS object with JS API naming ex: { className: "my-class", style: { fontFamily: "Verdana" }, id: "my-id"}
+      },
+      class: ''             // optional class names space delimited list for title item ex: "text-center"
+    },
+    {
+      name: 'Eventos',
+      url: '/actividades/consultar-eventos',
+      icon: 'icon-cursor'
+    },
+    {
+      title: true,
+      name: 'Perfil',
+      wrapper: {            // optional wrapper object
+        element: '',        // required valid HTML5 element tag
+        attributes: {}        // optional valid JS object with JS API naming ex: { className: "my-class", style: { fontFamily: "Verdana" }, id: "my-id"}
+      },
+      class: ''             // optional class names space delimited list for title item ex: "text-center"
+    },
+    {
+      name: 'Consultar perfil',
+      url: '/perfil',
+      icon: 'icon-pencil',
+    },
+  ],
+};
